+++ idf_component_tools/manifest/solved_component.py
@@ -4,8 +4,10 @@
 
 from idf_component_tools.serialization import serializable
 
+from ..constants import IDF_COMPONENT_STORAGE_URL
 from ..errors import LockError
 from ..manifest import ComponentRequirement, ComponentVersion
+from ..sources import WebServiceSource
 from ..sources.base import BaseSource
 
 try:
@@ -44,14 +46,26 @@ class SolvedComponent(object):
         return 'SolvedComponent <{}({}) {}>'.format(self.name, self.version, self.component_hash)
 
     def __str__(self):
-        return '{}({})'.format(self.name, self.version)
+        if (
+            isinstance(self.source, WebServiceSource)
+            and self.source._storage_url != IDF_COMPONENT_STORAGE_URL
+        ):
+            return '{name} ({version}) from {storage_url}'.format(
+                name=self.name,
+                version=self.version,
+                storage_url=self.source._storage_url,
+            )
+        else:
+            return '{name} ({version})'.format(name=self.name, version=self.version)
 
     @classmethod
     def fromdict(cls, details):
         try:
             source_details = dict(details['source'])
             source_name = source_details.pop('type')
-            source = BaseSource.fromdict(source_name, source_details)
+            if source_name == 'service' and 'storage_url' not in source_details:
+                source_details['storage_url'] = IDF_COMPONENT_STORAGE_URL
+            source = BaseSource.fromdict(source_name, source_details)[0]
             component_hash = details.get('component_hash', None)
             if source.component_hash_required and not component_hash:
                 raise LockError(