+++ tests/test_api_client.py
@@ -7,15 +7,24 @@ import pytest
 import vcr
 
 from idf_component_manager import version
-from idf_component_tools.api_client import APIClient, env_cache_time, join_url, user_agent
-from idf_component_tools.api_client_errors import NoRegistrySet
 from idf_component_tools.config import component_registry_url
 from idf_component_tools.constants import IDF_COMPONENT_REGISTRY_URL, IDF_COMPONENT_STORAGE_URL
+from idf_component_tools.registry.api_client import APIClient
+from idf_component_tools.registry.api_client_errors import NoRegistrySet
+from idf_component_tools.registry.base_client import env_cache_time, user_agent
+from idf_component_tools.registry.multi_storage_client import MultiStorageClient
+from idf_component_tools.registry.request_processor import join_url
+from idf_component_tools.registry.storage_client import StorageClient
 
 
 @pytest.fixture
 def base_url():
     return 'http://localhost:5000/api'
+
+
+@pytest.fixture
+def storage_url():
+    return 'http://localhost:9000/test-public'
 
 
 class TestAPIClient(object):
@@ -40,8 +49,8 @@ class TestAPIClient(object):
             assert join_url(*test['in']) == test['out']
 
     @vcr.use_cassette('tests/fixtures/vcr_cassettes/test_component_versions.yaml')
-    def test_version(self, base_url):
-        client = APIClient(base_url=base_url)
+    def test_version(self, storage_url):
+        client = StorageClient(storage_url=storage_url)
 
         # Also check case normalisation
         component = client.versions(component_name='Test/Cmp', spec='>=1.0.0')
@@ -50,20 +59,18 @@ class TestAPIClient(object):
         assert len(list(component.versions)) == 2
 
     @vcr.use_cassette('tests/fixtures/vcr_cassettes/test_component_details.yaml')
-    def test_component(self, base_url):
-        storage_name = 'http://localhost:9000/test-public/'
-
-        client = APIClient(base_url=base_url)
+    def test_component(self, storage_url):
+        client = StorageClient(storage_url=storage_url)
 
         # Also check case normalisation
         manifest = client.component(component_name='tesT/CMP')
 
         assert manifest.name == 'test/cmp'
         assert str(manifest.version) == '1.0.1'
-        assert manifest.download_url.startswith(storage_name)
-        assert manifest.documents['readme'].startswith(storage_name)
-        assert manifest.examples[0]['url'].startswith(storage_name)
-        assert manifest.license['url'].startswith(storage_name)
+        assert manifest.download_url.startswith(storage_url)
+        assert manifest.documents['readme'].startswith(storage_url)
+        assert manifest.examples[0]['url'].startswith(storage_url)
+        assert manifest.license_url.startswith(storage_url)
 
     def test_user_agent(self, base_url):
         ua = user_agent()
@@ -85,7 +92,7 @@ class TestAPIClient(object):
         'tests/fixtures/vcr_cassettes/test_api_cache.yaml',
         record_mode='none',
     )
-    def test_api_cache(self, base_url, monkeypatch, tmp_path):
+    def test_api_cache(self, storage_url, monkeypatch, tmp_path):
         """
         This test is checking api caching with using the same requests 2 times.
         In test_api_cache.yaml we have just one request, so if test is passed,
@@ -100,7 +107,7 @@ class TestAPIClient(object):
 
         monkeypatch.setenv('IDF_COMPONENT_API_CACHE_EXPIRATION_MINUTES', '180')
         monkeypatch.setenv('IDF_COMPONENT_CACHE_PATH', str(tmp_path))
-        client = APIClient(base_url=base_url)
+        client = StorageClient(storage_url=storage_url)
 
         client.component(component_name='test/cmp')
         client.component(component_name='test/cmp')
@@ -108,12 +115,12 @@ class TestAPIClient(object):
     @vcr.use_cassette('tests/fixtures/vcr_cassettes/test_api_information.yaml')
     def test_api_information(self, base_url):
         client = APIClient(base_url=base_url)
+        information = client.api_information()
+        assert information['components_base_url'] == 'http://localhost:9000/test-public'
 
-        assert client.storage_url == 'http://localhost:9000/test-public'
-
-    def test_file_adapter(self, base_url, fixtures_path):
+    def test_file_adapter(self, fixtures_path):
         storage_url = '{}{}'.format('file://', fixtures_path)
-        client = APIClient(base_url, storage_url=storage_url)
+        client = StorageClient(storage_url=storage_url)
 
         assert client.component(component_name='example/cmp').download_url == os.path.join(
             storage_url, '5390a837-5bc7-4564-b747-3adb22ad55f8.tgz'
@@ -122,8 +129,8 @@ class TestAPIClient(object):
     def test_no_registry_url_error(self, monkeypatch):
         monkeypatch.setenv('IDF_COMPONENT_STORAGE_URL', 'http://localhost:9000/test-public')
 
-        registry_url, storage_url = component_registry_url()
-        client = APIClient(base_url=registry_url, storage_url=storage_url, auth_token='test')
+        registry_url, _ = component_registry_url()
+        client = APIClient(base_url=registry_url, auth_token='test')
         with pytest.raises(NoRegistrySet):
             client.upload_version(component_name='example/cmp')
 
@@ -132,22 +139,22 @@ class TestAPIClient(object):
         monkeypatch.setenv('IDF_COMPONENT_REGISTRY_URL', '')
         monkeypatch.setenv('IDF_COMPONENT_API_TOKEN', '')
 
-        registry_url, storage_url = component_registry_url()
+        registry_url, storage_urls = component_registry_url()
         assert registry_url == IDF_COMPONENT_REGISTRY_URL + 'api/'
-        assert storage_url == IDF_COMPONENT_STORAGE_URL
+        assert storage_urls == [IDF_COMPONENT_STORAGE_URL]
 
     @vcr.use_cassette('tests/fixtures/vcr_cassettes/test_no_registry_url_use_static.yaml')
     def test_no_registry_url_use_static(self, monkeypatch):
         monkeypatch.setenv('IDF_COMPONENT_STORAGE_URL', 'http://localhost:9000/test-public')
 
-        registry_url, storage_url = component_registry_url()
-        client = APIClient(base_url=registry_url, storage_url=storage_url, auth_token='test')
+        registry_url, storage_urls = component_registry_url()
+        client = MultiStorageClient(storage_urls=storage_urls)
         client.component(component_name='espressif/cmp')  # no errors
 
     @vcr.use_cassette('tests/fixtures/vcr_cassettes/test_filter_yanked_version.yaml')
     @pytest.mark.parametrize('version', ['=1.1.0', '1.1.0', '==1.1.0,==1.1.0'])
-    def test_only_yanked_version_warning(self, base_url, version):
-        client = APIClient(base_url=base_url)
+    def test_only_yanked_version_warning(self, storage_url, version):
+        client = StorageClient(storage_url=storage_url)
 
         with pytest.warns(UserWarning, match='component you have selected has been yanked'):
             client.component(component_name='example/cmp_yanked', version=version)
@@ -163,8 +170,8 @@ class TestAPIClient(object):
             None,
         ],
     )
-    def test_filter_yanked_version(self, base_url, version):
-        client = APIClient(base_url=base_url)
+    def test_filter_yanked_version(self, storage_url, version):
+        client = StorageClient(storage_url=storage_url)
         result = client.component(component_name='example/cmp_yanked', version=version)
 
         assert result.version == '1.0.1'
@@ -188,3 +195,18 @@ class TestAPIClient(object):
         client = APIClient(base_url=base_url)
         with pytest.raises(Exception):
             client.token_information()
+
+    @vcr.use_cassette('tests/fixtures/vcr_cassettes/test_version_multiple_storages.yaml')
+    def test_version_multiple_storages(self, fixtures_path):
+        storage_file_path = 'file://{}/'.format(fixtures_path)
+        storage_urls = [storage_file_path, 'https://components-file.espressif.com']
+        client = MultiStorageClient(storage_urls=storage_urls)
+        result = client.versions(component_name='example/cmp')
+
+        assert result.versions
+        assert result.storage_url == storage_file_path
+
+        result = client.versions('espressif/mdns')
+
+        assert result.versions
+        assert result.storage_url == 'https://components-file.espressif.com'